Refactor database protocols, add filebase gateway
