+++ src/application/commands/register/handler.py
@@ -5,7 +5,7 @@ from uuid import uuid4
 from src.domain.models.user.model import User
 from src.domain.models.user.value_objects import UserId, Username
 from src.application.common.result import Result
-from src.application.common.passoword_encoder import PasswordEncoder
+from src.application.common.interfaces.passoword_encoder import PasswordEncoder
 from .command import RegisterCommand, RegisterCommandResult
 from .interfaces import RegisterCommandDBGateway
 from .errors import UsernameAlreadyExistsError
