+++ src/application/common/database_protocols/atomic.py
