+++ src/application/common/database_protocols/movie.py
