+++ src/application/common/database_protocols/user_movie_rating.py
