--- src/application/common/file_repository.py
@@ -1,10 +1,0 @@
-from abc import ABC, abstractmethod
-from typing import IO
-
-
-class FileRepostory(ABC):
-
-    @abstractmethod
-    def add_image(self, image: IO, key: str) -> None:
-        raise NotImplementedError
-
