+++ src/application/common/filebase_protocols/image.py
@@ -1,0 +1,10 @@
+from abc import abstractmethod
+from typing import Protocol
+
+
+class SupportsSaveImage(Protocol):
+
+    @abstractmethod
+    def add_image(self):
+        raise NotImplementedError
+
