+++ src/application/common/interfaces/database_gateway.py
@@ -1,21 +1,21 @@
 from typing import Protocol
 
-from .atomic import SupportsAtomic
-from .user import (
+from src.application.common.database_protocols.atomic import SupportsAtomic
+from src.application.common.database_protocols.user import (
     SupportsSaveUser,
     SupportsGetUserById,
     SupportsGetUserByUsername,
     SupportsCheckUsernameExistence,
     SupportsCheckUserIdExistence
 )
-from .movie import (
+from src.application.common.database_protocols.movie import (
     SupportsSaveMovie,
     SupportsGetMovieById,
     SupportsUpdateMovie,
     SupportsCheckMovieIdExistence,
     SupportsRemoveMovie
 )
-from .user_movie_rating import (
+from src.application.common.database_protocols.user_movie_rating import (
     SupportsSaveUserMovieRating,
     SupportsGetUserMovieRatingByUserIdAndMovieId,
     SupportsCheckUserMovieRatingExistence,
