+++ src/application/common/interfaces/filebase_gateway.py
@@ -1,0 +1,12 @@
+from typing import Protocol
+
+from src.application.common.filebase_protocols.image import (
+    SupportsSaveImage
+)
+
+
+class FilebaseGateway(
+    SupportsSaveImage,
+    Protocol
+):
+    ...
