+++ src/infrastructure/password_encoder.py
@@ -1,6 +1,6 @@
 from hashlib import sha256
 
-from src.application.common.passoword_encoder import PasswordEncoder
+from src.application.common.interfaces.passoword_encoder import PasswordEncoder
 
 
 class HashlibPasswordEncoder(PasswordEncoder):
